refactor: move encode page into library segment
