--- app/src/app/(dashboard)/@dialog/default.tsx
@@ -1,9 +1,0 @@
-type DialogDefaultProps = {
-  params: {
-    id: string
-  }
-}
-
-const DialogDefault: React.FC<DialogDefaultProps> = () => null
-
-export default DialogDefault
