+++ app/src/app/(libraries)/library/[slug]/encode/[id]/layout.tsx
@@ -1,0 +1,9 @@
+import { LayoutNarrow } from '@/components/layouts'
+
+type LibraryEncodeLayoutProps = React.PropsWithChildren
+
+const LibraryEncodeLayout: React.FC<LibraryEncodeLayoutProps> = ({ children }) => (
+  <LayoutNarrow>{children}</LayoutNarrow>
+)
+
+export default LibraryEncodeLayout
