+++ app/src/app/@dialog/default.tsx
@@ -1,0 +1,3 @@
+const DialogDefault: React.FC = () => null
+
+export default DialogDefault
